requirements.txt
requests>=2.31.0
trafilatura>=1.6.0
openai>=1.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0

